rename thumbnail_handler to yogo_result_handlers
